[WASM] Convert all ints to size_t

+++ tfjs-backend-wasm/src/cc/backend.cc
@@ -17,6 +17,7 @@
 #endif
 
 #include <xnnpack.h>
+#include <cstddef>
 #include <unordered_map>
 #include <utility>
 #include <vector>
@@ -26,28 +27,28 @@
 namespace {
 // Maps a unique tensor id to info about that tensor. The map owns all of its
 // entries.
-std::unordered_map<int, TensorInfo> data;
+std::unordered_map<size_t, TensorInfo> data;
 
 // Maps a tensor id to a vector of disposal functions registered on that tensor
 // id.
-std::unordered_map<int, std::vector<tfjs::backend::DisposeFunction>>
+std::unordered_map<size_t, std::vector<tfjs::backend::DisposeFunction>>
     disposal_callbacks;
 }  // namespace
 
 namespace tfjs {
 namespace backend {
-const TensorInfo &get_tensor_info(const int tensor_id) {
+const TensorInfo &get_tensor_info(const size_t tensor_id) {
   return data.at(tensor_id);
 }
 
-TensorInfo &get_tensor_info_out(const int tensor_id) {
+TensorInfo &get_tensor_info_out(const size_t tensor_id) {
   return data.at(tensor_id);
 }
 
-int xnn_operator_count = 0;
+size_t xnn_operator_count = 0;
 
 // Registers a disposal callback for a tensor id with a given callback function.
-void register_disposal_callback(const int tensor_id,
+void register_disposal_callback(const size_t tensor_id,
                                 const DisposeFunction dispose_fn) {
   if (disposal_callbacks.count(tensor_id) == 0) {
     // We move callbacks to avoid a copy.
@@ -58,7 +59,7 @@ void register_disposal_callback(const int tensor_id,
   }
 }
 
-const int num_tensors() { return data.size(); }
+const size_t num_tensors() { return data.size(); }
 
 }  // namespace backend
 
@@ -74,14 +75,15 @@ void init() { xnn_initialize(nullptr); }
 #ifdef __EMSCRIPTEN__
 EMSCRIPTEN_KEEPALIVE
 #endif
-void register_tensor(const int tensor_id, const int size, void *memory_offset) {
+void register_tensor(const size_t tensor_id, const size_t size,
+                     void *memory_offset) {
   data.emplace(tensor_id, TensorInfo{memory_offset, size});
 }
 
 #ifdef __EMSCRIPTEN__
 EMSCRIPTEN_KEEPALIVE
 #endif
-void dispose_data(const int tensor_id) {
+void dispose_data(const size_t tensor_id) {
   data.erase(tensor_id);
 
   // Call all disposal callbacks for this tensor id.
@@ -102,7 +104,7 @@ EMSCRIPTEN_KEEPALIVE
 void dispose() {
   // We have to create a separate vector of tensor ids because we erase from the
   // map while we're iterating it.
-  std::vector<int> tensor_ids_to_dispose;
+  std::vector<size_t> tensor_ids_to_dispose;
   for (const auto &element : data) {
     tensor_ids_to_dispose.push_back(element.first);
   }
