+++ tfjs-backend-wasm/src/cc/kernels/BatchMatMul.cc
@@ -16,10 +16,12 @@
 #include <emscripten.h>
 #endif
 
+#include <cstddef>
+
 #include "src/cc/backend.h"
 #include "src/cc/util.h"
 
-const int kBlockSize = 48;
+const size_t kBlockSize = 48;
 
 namespace tfjs {
 namespace wasm {
@@ -29,12 +31,12 @@ extern "C" {
 #ifdef __EMSCRIPTEN__
 EMSCRIPTEN_KEEPALIVE
 #endif
-void BatchMatMul(const int a_id, const int b_id, const int shared_dim,
-                 const int left_dim, const int right_dim, const int batch_dim,
-                 const int a_batch, const int a_outer_step,
-                 const int a_inner_step, const int b_batch,
-                 const int b_outer_step, const int b_inner_step,
-                 const int out_id) {
+void BatchMatMul(const size_t a_id, const size_t b_id, const size_t shared_dim,
+                 const size_t left_dim, const size_t right_dim,
+                 const size_t batch_dim, const size_t a_batch,
+                 const size_t a_outer_step, const size_t a_inner_step,
+                 const size_t b_batch, const size_t b_outer_step,
+                 const size_t b_inner_step, const size_t out_id) {
   auto& a_info = backend::get_tensor_info(a_id);
   auto& b_info = backend::get_tensor_info(b_id);
   auto& out_info = backend::get_tensor_info_out(out_id);
@@ -43,25 +45,25 @@ void BatchMatMul(const int a_id, const int b_id, const int shared_dim,
   const float* b_buf = b_info.f32();
   float* out_buf = out_info.f32_write();
 
-  const int size = left_dim * right_dim;
+  const size_t size = left_dim * right_dim;
 
   // Zero out the output buffer because it might have been used before.
   std::fill(out_buf, out_buf + batch_dim * size, 0);
 
-  for (int b = 0; b < batch_dim; ++b) {
-    for (int i0 = 0; i0 < left_dim; i0 += kBlockSize) {
-      for (int j0 = 0; j0 < right_dim; j0 += kBlockSize) {
-        for (int k0 = 0; k0 < shared_dim; k0 += kBlockSize) {
+  for (size_t b = 0; b < batch_dim; ++b) {
+    for (size_t i0 = 0; i0 < left_dim; i0 += kBlockSize) {
+      for (size_t j0 = 0; j0 < right_dim; j0 += kBlockSize) {
+        for (size_t k0 = 0; k0 < shared_dim; k0 += kBlockSize) {
           // for when kBlockSize doesn't evenly divide the input
-          const int i_block = std::min(i0 + kBlockSize, left_dim);
-          const int j_block = std::min(j0 + kBlockSize, right_dim);
-          const int k_block = std::min(k0 + kBlockSize, shared_dim);
+          const size_t i_block = std::min(i0 + kBlockSize, left_dim);
+          const size_t j_block = std::min(j0 + kBlockSize, right_dim);
+          const size_t k_block = std::min(k0 + kBlockSize, shared_dim);
 
-          for (int i = i0; i < i_block; ++i) {
-            for (int j = j0; j < j_block; ++j) {
+          for (size_t i = i0; i < i_block; ++i) {
+            for (size_t j = j0; j < j_block; ++j) {
               float sum = 0.0;
 
-              for (int k = k0; k < k_block; ++k) {
+              for (size_t k = k0; k < k_block; ++k) {
                 sum +=
                     a_buf[b * a_batch + i * a_outer_step + k * a_inner_step] *
                     b_buf[k * b_inner_step + j * b_outer_step + b * b_batch];
