+++ tfjs-backend-wasm/src/cc/kernels/FloorDiv.cc
@@ -17,6 +17,7 @@
 #endif
 
 #include <cmath>
+#include <cstddef>
 
 #include "src/cc/backend.h"
 #include "src/cc/binary.h"
@@ -30,9 +31,10 @@ extern "C" {
 #ifdef __EMSCRIPTEN__
 EMSCRIPTEN_KEEPALIVE
 #endif
-void FloorDiv(const int a_id, const size_t* a_shape_ptr, const int a_shape_len,
-              const int b_id, const size_t* b_shape_ptr, const int b_shape_len,
-              const DType dtype, const int out_id) {
+void FloorDiv(const size_t a_id, const size_t* a_shape_ptr,
+              const size_t a_shape_len, const size_t b_id,
+              const size_t* b_shape_ptr, const size_t b_shape_len,
+              const DType dtype, const size_t out_id) {
   auto& a_info = backend::get_tensor_info(a_id);
   switch (dtype) {
     case DType::float32:
@@ -40,8 +42,8 @@ void FloorDiv(const int a_id, const size_t* a_shape_ptr, const int a_shape_len,
                  [](float a, float b) { return floor(a / b); });
       break;
     case DType::int32:
-      binary_i32(a_id, b_id, out_id, [](int a, int b) {
-        return static_cast<int>(floor(static_cast<float>(a) / b));
+      binary_i32(a_id, b_id, out_id, [](size_t a, size_t b) {
+        return static_cast<size_t>(floor(static_cast<float>(a) / b));
       });
       break;
     default:
