+++ tfjs-backend-wasm/src/cc/kernels/Square.cc
@@ -16,6 +16,8 @@
 #include <emscripten.h>
 #endif
 
+#include <cstddef>
+
 #include "src/cc/backend.h"
 #include "src/cc/unary.h"
 
@@ -31,7 +33,9 @@ extern "C" {
 #ifdef __EMSCRIPTEN__
 EMSCRIPTEN_KEEPALIVE
 #endif
-void Square(const int x_id, const int out_id) { unary(x_id, out_id, square); }
+void Square(const size_t x_id, const size_t out_id) {
+  unary(x_id, out_id, square);
+}
 
 }  // extern "C"
 }  // namespace wasm
